+++ NSArrayExtended.m
@@ -1,7 +1,7 @@
 //  NSArrayExtended.m
 //  LaTeXiT
 //  Created by Pierre Chatelier on 4/05/05.
-//  Copyright 2005-2018 Pierre Chatelier. All rights reserved.
+//  Copyright 2005-2019 Pierre Chatelier. All rights reserved.
 
 // This file is an extension of the NSArray class
 
@@ -71,7 +71,7 @@
 }
 //end reversed
 
--(NSArray*) arrayByAddingObject:(id)object atIndex:(unsigned int)index
+-(NSArray*) arrayByAddingObject:(id)object atIndex:(NSUInteger)index
 {
   #ifdef ARC_ENABLED
   NSMutableArray* result = [self mutableCopy];
@@ -83,7 +83,7 @@
 }
 //end arrayByAddingObject:atIndex:
 
--(NSArray*) arrayByMovingObjectsAtIndices:(NSIndexSet*)indices toIndex:(unsigned int)index
+-(NSArray*) arrayByMovingObjectsAtIndices:(NSIndexSet*)indices toIndex:(NSUInteger)index
 {
   #ifdef ARC_ENABLED
   NSMutableArray* result = [self mutableCopy];
