+++ NSMenuExtended.h
@@ -3,7 +3,7 @@
 //  LaTeXiT
 //
 //  Created by Pierre Chatelier on 21/05/10.
-//  Copyright 2005-2018 Pierre Chatelier. All rights reserved.
+//  Copyright 2005-2019 Pierre Chatelier. All rights reserved.
 //
 
 #import <Cocoa/Cocoa.h>
@@ -12,7 +12,7 @@
 @interface NSMenu (Extended)
 
 -(NSMenuItem*) addItemWithTitle:(NSString*)aString target:(id)target action:(SEL)aSelector
-                  keyEquivalent:(NSString*)keyEquivalent keyEquivalentModifierMask:(int)keyEquivalentModifierMask
-                  tag:(int)tag;
+                  keyEquivalent:(NSString*)keyEquivalent keyEquivalentModifierMask:(NSInteger)keyEquivalentModifierMask
+                  tag:(NSInteger)tag;
 
 @end
